#4489 Improved re-rendering of content toolbar for geostory (#4490)
