+++ web/client/components/geostory/contents/__tests__/Column-test.jsx
@@ -41,27 +41,34 @@ describe('Column component', () => {
         const el = container.querySelector('.add-bar');
         expect(el).toExist();
     });
-    it('Column contents has proper toolbars', () => {
-        // text content should contain only delete button
-        ReactDOM.render(<Column mode={Modes.EDIT} contents={[{ type: ContentTypes.TEXT, html: '<p id="TEST_HTML">something</p>' }]} />, document.getElementById("container"));
-        const textToolbar = document.querySelector('.ms-content-toolbar .btn-group');
-        expect(textToolbar).toExist();
-        expect(textToolbar.querySelectorAll('button').length).toBe(1);
-        expect(textToolbar.querySelector('button .glyphicon-trash')).toExist(); // align tool
+    describe('Column contents has proper toolbars', () => {
+        it('text', () => {
+            // text content should contain only delete button
+            ReactDOM.render(<Column mode={Modes.EDIT} contents={[{ type: ContentTypes.TEXT, html: '<p id="TEST_HTML">something</p>' }]} />, document.getElementById("container"));
+            const textToolbar = document.querySelector('.ms-content-toolbar .btn-group');
+            expect(textToolbar).toExist();
+            expect(textToolbar.querySelectorAll('button').length).toBe(1);
+            expect(textToolbar.querySelector('button .glyphicon-trash')).toExist(); // align tool
+        });
+        it('media', () => {
+            // media and image contents must have edit, resize and align tools
+            ReactDOM.render(<Column mode={Modes.EDIT} contents={[{ type: ContentTypes.MEDIA }]} />, document.getElementById("container"));
+            let mediaToolbar = document.querySelector('.ms-content-toolbar .btn-group');
+            expect(mediaToolbar).toExist();
 
-        // media and image contents must have edit, resize and align tools
-        ReactDOM.render(<Column mode={Modes.EDIT} contents={[{ type: ContentTypes.MEDIA }]} />, document.getElementById("container"));
-        let mediaToolbar = document.querySelector('.ms-content-toolbar .btn-group');
-        expect(mediaToolbar).toExist();
-        // TODO:  check (empty) media button (Should be type: media)
-        // image contents must have edit, resize align and delete tools
-        ReactDOM.render(<Column mode={Modes.EDIT} contents={[{ type: MediaTypes.IMAGE }]} />, document.getElementById("container"));
-        mediaToolbar = document.querySelector('.ms-content-toolbar .btn-group');
-        expect(mediaToolbar).toExist();
-        expect(mediaToolbar.querySelectorAll('button').length).toBe(4);
-        expect(mediaToolbar.querySelector('button .glyphicon-pencil')).toExist(); // edit tool
-        expect(mediaToolbar.querySelector('button .glyphicon-resize-horizontal')).toExist(); // resize tool
-        expect(mediaToolbar.querySelector('button .glyphicon-align-center')).toExist(); // align tool
-        expect(mediaToolbar.querySelector('button .glyphicon-trash')).toExist(); // delete tool
+        });
+        it('image', () => {
+            // TODO:  check (empty) media button (Should be type: media)
+            // image contents must have edit, resize align and delete tools
+            ReactDOM.render(<Column mode={Modes.EDIT} contents={[{ type: MediaTypes.IMAGE }]} />, document.getElementById("container"));
+            let mediaToolbar = document.querySelector('.ms-content-toolbar .btn-group');
+            expect(mediaToolbar).toExist();
+            const buttons = mediaToolbar.querySelectorAll('.ms-content-toolbar .btn-group button');
+            expect(buttons.length).toBe(4);
+            expect(mediaToolbar.querySelector('button .glyphicon-pencil')).toExist(); // edit tool
+            expect(mediaToolbar.querySelector('button .glyphicon-resize-horizontal')).toExist(); // resize tool
+            expect(mediaToolbar.querySelector('button .glyphicon-align-center')).toExist(); // align tool
+            expect(mediaToolbar.querySelector('button .glyphicon-trash')).toExist(); // delete tool
+        });
     });
 });
